Refactor schema.py
