+++ photo/mutations.py
@@ -1,0 +1,6 @@
+import strawberry
+
+
+@strawberry.type
+class Mutation:
+    pass
